Set main text to generate a new language when clicked
Set the "Today I'll code in" text on the webpage to change the random language when clicked. Also separated the code to set the language name displayed into its own function, so it can be called repeatedly.

Also removed the dummy text in the random language section, so that it doesn't display as the page is loading.